+++ CHANGELOG.md
@@ -56,6 +56,12 @@ Main (unreleased)
  
 - Add queueing logs remote write client for `loki.write` when WAL is enabled. (@thepalbi)
 
+- New Grafana Agent Flow components:
+
+  - `otelcol.processor.filter` - filters OTLP telemetry data using OpenTelemetry
+    Transformation Language (OTTL). (@hainenber)
+
+
 ### Bugfixes
 
 - Fixed an issue where `loki.process` validation for stage `metric.counter` was